[#144016539] Fix python packaging.
We need all the files we're going to distribute (the contents of `src`) to be in a directory named after the package, because this directory will land in `site-packages` - without it, we lose the package namespace. Set up this directory in the deploy script so that `python setup.py sdist` will package it up, and use the manifest to exclude everything else. Make sure to add an __init__.py to svGhostDriver, because otherwise `python setup.py install` will silently ignore it.