[#147678299] Don't trust QT error codes.
We thought QT would preserve HTTP error codes but in fact it does strange conversions, e.g. 404 becomes 203. So we can't tell easily that an `onResourceError` response is an HTTP error by checking its status code. Instead, ignore such responses whenever we already have a status code in the HAR entry, because that means the error must be HTTP.